# sectok e2e faster