@tsachiherman 2021-05-26
how does this solution handle error code generated by goal ? would it still fail the script as before ?