@brianolson 2021-05-26
bash `wait` will now return the error code of the awaited background process.
```sh
#!/usr/bin/env bash

set -ex
set -o pipefail
export SHELLOPTS

false &
WA=$!
wait $WA
echo still going
`
above script does not reach 'still going'